Set TTL only for files which match prospector (#2836)
Previously each prospector reset the TTL for all states which could mean that if one prospector has a different clean_inactive value then an other, the last value would win on startup. This only has an affect in case filebeat is restarted. Now the state is only updated in case the state.Source is part of the prospector glob. To make sure all states are still set to Finished on the registry site, this is done for all states.

(cherry picked from commit d63d2f59ca0332220b79bdc1356430bfff8ca610)