Show all missing plugins in the same err message (#3706)
A drawback of the error handling in #3676 was that if more than one
plugin is missing in ES, the error only reported the first one. This means
that the user might go through an annoying trial and error cycle.

To solve this, we make the modules declare their processor requirements
in the `manifest.yml` file and we compare them with the available processors
from calling `/_nodes/ingest`.

In case the module author doesn't declare the required processor plugins, the
error handling in #3676 still applies.