Fix issue that new prospector was not reloaded on conflict (#4128) (#4227)
A new prospector cannot be started if one of the states he starts with is not set to Finished. In this case the prospector should wait until the old prospector is shutdown and try again on the next reload. This did not work as intended so far because if loading failed, the running prospector was stopped and the non running was not started again on the next run. This change should fix it.

See also https://discuss.elastic.co/t/error-when-reloading-prospectors-for-filebeat/83082/5

Closes https://github.com/elastic/beats/issues/4133
(cherry picked from commit 325c9c84a876b8e309c2c720d14249fed249af98)