Cherry-pick #4506 to 5.x: Filebeat modules: Machine learning jobs (#4588)
* Filebeat modules: Machine Learning jobs (#4506)

This adds support for loading ML configurations (job + datafeed) from the filebeat modules.
An example ML configuration is added to the Nginx Filebeat module. This sample applies
ML anomaly detection on the response codes.

The loading is implemented as part of the `setup` command and part of the `--setup` flag.

If a job configuration with the same ID exists, it is not overwritten, because deleting jobs
could potentially delete user data. The user should manually delete the jobs in the UI if they
want to upgrade.
(cherry picked from commit ae4fed5ff4a84d6a1b78c1861b1ad5935ba7d5db)

* use 5.5 snapshots for now

* Adjusted test for 5.x

* Another test adjustments

* On a 400 response on _xpack, assume that Xpack is not installed

* Another test fix