Do not require template if index change and template disabled (#5319) (#5334)
* Do not require template if index change and template disabled

If the elasticsearch index name is changed, its required to set a template. The error message was also shown when template itself are disabled. This is now changed that if templates are disabled, no error is returned.

Closes https://github.com/elastic/beats/issues/5308

* Add Changelog

(cherry picked from commit b56accf535ca3705595aa57ec5a752a7394a91cb)