Improve test stability (#5567) (#5594)
Currently CI is constantly failing because of flaky Kafka and sometimes Kibana tests. This skips the Kafka tests for now and increases the timeout for the Kibana tests. Same is done for Dropwizard. The timeout for the locks was increase to see if it helps to make the builds more stable. This should make our CI builds useful again and not constantly red.

For Kafka a follow up PR is here to remove the test skipping again and make it more stable.

(cherry picked from commit 922ee2ebd137ce72d04222e35f6c6f2acb9abdf7)