Fix default host to localhost for Kibana dashboard loading (#5769) (#5781)
On my machine when loading dashboards with `metricbeat setup dashboards` it tried to load the dashboards to `ruflin:5601` as my host name is `ruflin`. But it should load it to localhost by default.

The identification of the hostname / beat.name was added to replace it in the host dashboards with the actual host name. This was also applied to the hostname of the instance it should be loaded to. I think this was unintentional. This PR sets the default to localhost.

(cherry picked from commit 03037a8e411e932e6b485324e660c10eb0fa96a6)