Fix flaky clean_removed test (#8092) (#8118)
This is a potential fix for the flaky clean_removed test. The problem in the test seems to be that sometimes not all states are cleaned up yet. This is changing it by waiting for all pending cleanups to happen.

Closes https://github.com/elastic/beats/issues/7690

(cherry picked from commit 66258350a3cede9a128426fae5dfb02e1cc90adb)