Cherry-pick #8346 to 6.4: Correctly parse rfc3339 timezone and Nanosecond (#8432)
Cherry-pick of PR #8346 to 6.4 branch. Original message: 

The code was not correctly creating the time and was isntead making
events appear to be a few years in the future. The nanosecond extraction
was not on the right scale for the time.Date, the commit fixes the
behavior and do assertions with the go parser.

bug was reported by @vjsamuel