Remove tls certificates from heartbeat data (#8390) (#8477)
This was recently added (not yet released) for the 6.5 release. Having lived with it for a bit, I think we
should not include the full PEM certs on each check. Heartbeat checks can be quite frequent, and a cert chain
can be quite large.

I think we should hold off here, and if there's demand from the community add it in later.

(cherry picked from commit 55c4950a2a20140d6a70cf069083977018e7a7d5)