Get cluster UUID from Info API call (#8592) (#8610)
Starting from 6.5.0, Elasticsearch [will return](https://github.com/elastic/elasticsearch/pull/32206) the `cluster_uuid` as part of the `GET _cluster/stats` API response. However, the Elasticsearch metricbeat module has been in existence since 6.3.0 so it must fetch the cluster UUID from the `GET /` Elasticsearch API.

This PR makes it so the `GET /` API is always called and the cluster UUID from it is used. This is obviously not ideal in terms of API calls, but it's the simplest fix (for now, until we can figure out a nicer way to deal with version differences).

(cherry picked from commit 403cec011864e6007c4c2674abd5f32532898c92)