Stop recommending publish debugging when running beats. (#8935) (#9019)
Our instructions for running beats on every platform but OSX describe starting
them as a managed service. Mysteriously, for OSX we run the beat in the forgeground
and invoke it with `-d "publish"`. This leads to really verbose output that
is really unnecessary for someone trying a beat for the first time.

My suspicion is the `d` flag made it in there in a mistake and was perpetuated
through no intentional plan.

(cherry picked from commit 4a423480da67674687ab9b693771e24e184b6a61)