Fix alias field generation in docs (#9269) (#9498)
Alias fields documented in the fields list were not correct. The reason is that `path` was used in our generation of the docs for a different puprose. This one was renamed and reporting of path was added for field alias.

(cherry picked from commit 6a2ae3383266bfcd0d825a6fb623604215a62463)