Fix flaky ml test by increasing timeout (#9523) (#9527)
The ml tests sometimes fail on Travis or Jenkins but it seems when looking at the output of the files normally all entries are there. As they pass most of the time I assume it is an issue related to the timeout. Increasing the timeout to 60s.

(cherry picked from commit 29e9395f60bf8e90f9a76401d283d4dceb93b161)