Fix issue with ILM that write index was created before template was loaded (#9617) (#9628)
The write index for ILM was created before the template was loaded. This meant the first template did not contain all the mappings and policy for ILM so was never rolled over. This change makes sure the template is loaded for ILM is setup up and adds a test to confirm it.

Further ILM was not respected when running `metricbeat setup --template`. This is now changed too.



(cherry picked from commit e833684912934150ef129b8a8608d4517aa83b66)