Cherry-pick #9642 to 6.x: Export field for xpack.enabled setting (#9655)
Cherry-pick of PR #9642 to 6.x branch. Original message: 

The cleanup in https://github.com/elastic/beats/pull/8308 introduced a regression which caused the `xpack.enabled` flag in the `kibana` module to stop having any effect.

This PR fixes that issue by reverting the troublesome bits from #8308, ensuring that the struct field for the parsed configuration setting is exported.