Backport: Generate fields.go for netflow (#9740)
* Backport fields generation code

Cherry-pick select parts of f54df4f36c2cdcaccf677d05bb420ff06cb1d1a7

* Generate fields.go for netflow input

* Generate fields.go for netflow input

Update x-pack/filebeat to generate a fields.go for each input type. And update the code that
generates the include/list.go to import each input.

* Fix netflow exporter version datatype

It was "int" which is not a valid type. Changed it to long rather than integer because we use that everywhere for numbers.

Cherry-pick parts of c44eb1b88c29919b47f88cdc331b997b8aa9546f (#9686)

* Update x-pack/filebeat to use the xpack.mk

The existing Filebeat Makefile could not be reused because generate_imports.py is not portable.

* Add clean and exportDashboard targets

* createDir in GenerateModuleReferenceConfig

* Remove makefile hack to test x-pack/filebeat