[Auditbeat] Report process errors (#9693) (#9844)
Changes the process metricset to keep iterating through processes even when an unexpected error occurs. The error will be stored in the Process object and sent to Elasticsearch as well as logged as a warning. This only happens the first time the error is encountered for a process, not on subsequent collection cycles.

(cherry picked from commit f7ce3b1a6b52e03aeba3c539c9f295eb12d8aa9e)