[Auditbeat] Cherry-pick #9327 to 6.x: Login metricset (#10509)
Adds the login metricset to the Auditbeat system module as the last of the six initial metricsets. It only works on Linux, and detects not just user logins and logouts, but also system boots and shutdowns.

It works by reading the /var/log/wtmp and /var/log/btmp file (and rotated files) present on Linux systems. In reading a file, it is similar to Filebeat, except that UTMP is a binary format, so reading happens using a binary Go reader.

(cherry picked from commit 1cba1d6b6293401a4442b5bb058f0281fe61b9d9)