[Auditbeat] System module: Fix and unify bucket closing logic (#10897) (#11027)
The `host` dataset is erroneously trying to save state in its `Close()` method. It should have saved the state earlier - usually at the end of `Fetch()` - and then should only close the bucket (something it is not doing at all). At the same time, it is not saving state in its `reportState()` method. Combined, this can lead to an error when the dataset is terminated before the first regular `reportChanges()` is run.

This fixes both issues and furthermore unifies the bucket closing logic across all six datasets of the System module.

(cherry picked from commit 706b7fc6c15544b7c7cbbfc84f5b358143723513)