Don't generate autodiscover config when no port matches host hints (#12086) (#12098)
On metricbeat, when the host autodiscover hint is used, and it includes
the port, one of the exposed ports has to match with the one in the
hint. If not, no configuration should be generated. If it is generated,
it will have empty hosts, what would lead to unexpected errors as the
seen in #8264.

(cherry picked from commit ba36dce6d2a4a577bda3b45823cc847e80d5c2b5)