[Auditbeat] Cherry-pick #12591 to 7.2: Host: Fix reboot detection logic (#12594)
On Windows, `BootTime` is not fully accurate and can vary by a few milliseconds. This causes a lot of false positive `event.action: reboot` events.

This changes to only report a reboot if the new `BootTime` is at least 1 second after the old. This should fix Windows and not affect the other platforms, assuming it's impossible to reboot a system twice in 1 second.

(cherry picked from commit 4c4482827ac06d6448fa7465a988467daf55d904)