[Auditbeat] Cherry-pick #12940 to 6.8: Export Host fields to g… (#13033)
In the `system/host` dataset, the fields of the `Host` struct are not exported, and so cannot be picked up by `gob`. This prevents it from being written to disk to detect changes between restarts and causes error messages. This changes the fields to be exported.

(cherry picked from commit c476be4b6af287546cc3017bbfc13f8fcdcf162e)