Cherry-pick #13812 to 7.4: Allow users to select the cloud met… (#13815)
* Allow users to select the cloud metadata providers (#13812)

* Allow users to select the cloud metadata providers

We introduce a new setting 'providers' to the add_cloud_metadata
processor.

By now all the implementation for metadata providers requires developers
to mark a provider as 'local'. The alibaba and tencent providers are not
marked as local by now.

If the 'providers' setting is not used, then no all providers marked as
'local' are applied. This is a breaking change, because alibaba and
tencent providers will not be enabled anymore by default.

If the providers setting is used, only the selected providers will be
used.

(cherry picked from commit fa6d344a299630c2727895834472f2c4490079d5)