[Agent] Fix flaky tests concerning the TockenBucket (#15282)
This PR changes how we do testing for the bucket, instead of using time
based assertions we use a scheduler. In the tests we can manually
trigger the tick when the TokenBucket can drop. We also rely on the
blocking behavior and the fact that the tests will timeout.

So errors like the following in the CI should be removed:

```
=== RUN   TestTokenBucket

--- FAIL: TestTokenBucket (0.22s)

    token_bucket_test.go:37:

        	Error Trace:	token_bucket_test.go:37

        	Error:      	Not equal:

        	            	expected: 4

        	            	actual  : 5

        	Test:       	TestTokenBucket
```