Cherry-pick: Migrate scripts to python 3 (#14798) (#16302)
Several changes in test and tools code to migrate to Python 3 and stop
supporting Python 2. From now on, Python 3 is required to run tests and
many make and mage targets.

Switch over to `python3 -m venv` instead of `virtualenv`. It's the
recommended way to create virtual Python environments in 3.4 and above.

Remove any direct references to python to always allow a specific python
binary or version to be used with the PYTHON_EXE environmet variable.

Use python3 in the shebang for python scripts. python is going to be
reserved for python2 in Ubuntu, and PEP394 mentions that python should be
used in the shebang line only for scripts that are source compatible with
both Python 2 and 3.

Update documentation to reflect these changes and add information about
the use of Python for beats development.

Prepare the tests for checks based on the local ip, this IP can be
[::1] if ipv4 is not available or 127.0.1.1 in some machines, as in the
newer Travis images.

Change checks based on sys.platform, as the values for linux have changed
in Python 3.

Several assertions in tests have been adapted to its newer versions.

Some flaky tests that are more flaky with Python 3 have been fixed or skipped.

Replace use of nosetests with mage wrapper in script for Windows CI workers.
Mage manages their own python virtual environments, this is preferred to directly
call python commands.

(cherry picked from 05cadb99b49a1b2f520ecb75cf8dff17253c6267)

Co-authored-by: Noémi Ványi <kvch@users.noreply.github.com>
Co-authored-by: Benjamin Wohlwend <bw@piquadrat.ch>
Co-authored-by: Michael Madden <mikemadden42@users.noreply.github.com>
Co-authored-by: Andrew Kroh <andrew.kroh@elastic.co>