Fix data race in convert processor (#17032) (#17045)
If the convert processor was used in the global context it could lead to data races because there was a
variable that was reused across executions. When processors are used in the global context they are
shared across individual publisher clients so you could end up with a data race.

The fix here was to replace the shared state with a local variable. In testing this didn't make much of
a difference in the number of allocations.

(cherry picked from commit d6fa249cf6968049418d225c7e971a54b230b011)