Cherry-pick #17028 to 7.x: Ensure pprof endpoint is listening on startup (#17081)
pprof HTTP endpoint was being initialized asynchronously. In tests we
have found that sometimes after Metricbeat has started, it fails to
connect to the pprof endpoint. This can happen if the server is not
listening before the rest of the beat is initialized.

This change ensures that the listener for this endpoint is started
before continuing with the rest of the initialization.

Some comments and error messages have been polished.

(cherry picked from commit fab9a12315b8aa8e0363d4a3800f451d40d9cd35)