Install all packages in mage (#17388) (#17436)
This PR fixes the install command of `mage`. With the adoption of go modules adding `/...` has become necessary in order to install the pacakge.

This issue prevents beats from installing `mage` if it is not present. One of the symptoms of this problem is the failure to generate new Beats if `mage` is not available.

(cherry picked from commit 54cdaddf5734bdbe33d0aec0a49e64b2c341b073)