Add data directories to kubernetes manifests (#17429) (#17508)
Mount data directories in Metricbeat and Auditbeat pods too. They were
defined in some places but not mounted.

For Auditbeat this is important so it doesn't have to rebuild its file
integrity database on each run. For Metricbeat I don't think it is so
important, but adding it just in case for consistency with the other
reference manifests.

(cherry picked from commit 168d8ebbdaa59a6bef19fafb1a576e2901e01bde)