Update running-on-kubernetes.asciidoc
Support Ticket #00516669 - Customer had problems getting Kubernetes working correctly, had to use the master branch release. I wanted to update the docs to let others know about this work around as well as to ensure that their indices are generated by the same version of MB that's running. This is my first docs PR so not sure if this is the right location or wording for that information. Related to https://github.com/elastic/content/issues/4174