[Agent] Enable monitoring by default (#18225)
This enables monitoring of the agent by default. This PR only changes it in the configuration file. If we decide to move forward with this default, it should also be changed in the code.

The reason I think monitoring should be enabled out of the box as it will make debugging agents easier as all the data is directly avaiable in Elasticsearch and enhances the out of the box experience.

(cherry picked from commit ea39b3ccd529d7329306d2a26821be568a281bb5)