Cherry-pick #18991 to 7.8: Fix translate_sid's empty target field handling (#19084)
The translate_sid processor only requires one of the three target fields to be configured. It should work properly when some of the targets are not set, but it doesn't check if the are empty strings. So it ends up adding target fields that are empty strings to the event (e.g. "": "Group").

Closes #18990

(cherry picked from commit 4b19f09e7515a4cd814b3b0306d6f3fd066b12b6)