Disable the option of running setup --machine-learning on its own (#20241) (#20266)
This PR disables running `setup --machine-learning` on its own. From now on `setup --machine-learning --dashboards` has to be specified to enable ML jobs.

ML jobs require an index template pattern in Kibana. If it is not configured, loading the jobs fails. However, loading this asset is done by the dashboards step of setup.

Loading index template patterns are too tightly coupled with Kibana and dashboards and it needs refactoring. As this code is way too complex and going to be removed soon, I opted for a workaround instead of a fix.

Closes #19964
Closes #12998
Closes #20022
(cherry picked from commit 4ebf884034aa14556cadd0be94d5aa92ddd631e6)