[Elastic Agent] Fix passing of enrollment token to Docker container (#23155)
In 7.10.* and 7.x if an enrollment token is passed to the Docker container but no setup is used, the elastic-agent prints out the following error:

```
Error: accepts 2 arg(s), received 1
Usage:
  elastic-agent enroll <kibana_url> <enrollment_token> [flags]
```

The reason for this is that the apikey variable is overwritten with an empty value if no setup is used. It seems a change was made where the apikey overwrite landed outside the if clause instead of inside. Interestingly the bug does not exist in master.

This change needs to be backported to 7.10. too.