Prevent log input from sending duplicate messages due to file renaming (#21911) (#23743)
Input:Log: Reset TTL of registry state when a file is renamed. In some
rare cases the registry state is marked for removal (TTL is set to 0)
while the file is only renamed. Log detects the renaming of the file
and updates the name of the file. As the file still exists it should
also update the TTL of the renamed file.

(cherry picked from commit 7d6c6fe1b11af8fb5c3e3a67930db0b3918cf00c)

Co-authored-by: William Deurwaarder <william.git@xs4all.nl>