[Heartbeat] Report first synth error on summary (#24194) (#24222)
Fixes https://github.com/elastic/kibana/issues/92066

Prior to this for the summary.error field for browser events we'd
report the last encountered error. We now report the first. This
prevents the issue where on a failing run the synthetics library
considers an exit status of 1 to be an error. This will always be the
case when a step fails, meaning that the last error will always be a
status code error stepping on the actual step error.

(cherry picked from commit c5821d99bccea54f8253f3df247db3be4658834d)