[Heartbeat] Add sandbox options, defaulted to off (#24172) (#24358)
Fixes #22901

Adds support for the sandbox options in synthetics. I haven't added tests here because we really need E2E tests for this, but those live elsewhere. Given that we're still in experimental, I think it's OK to merge this feature as-is, and circle back to it later if you're OK with it @blakerouse , I can make a placeholder issue.

(cherry picked from commit c9d6b0041454df15b16e147a4f33fa1a5fcea404)