Cherry-pick #24719 to 7.12: [Filebeat] Fix gcp/vpcflow module defaulting to file input (#24734)
The pubsub input was renamed from google-pubsub to gcp-pubsub and the vpcflow filesets manifest
was not updated. As a result it's defaulting to the file input. This fixes the manifest.

The workaround is to explicitly set the input type:

    vpcflow:
      enabled: true
      var.input: gcp-pubsub

(cherry picked from commit c5ef8ed9a3920a02f293b49e0e20458186ee63ef)