Cherry-pick #24456 to 7.x: Osquerybeat (#24947)
Cherry-pick of PR #24456 to 7.x branch. Original message: 

## What does this PR do?

This PR is the first cut of the agent actions with osquerybeat. This allows agent to run osquerybeat, dispatch the ad-hoc queries received with "agent actions" and send the results back.
The osquerybeat was migrated from https://github.com/elastic/osquerybeat, the original POC we worked on awhile back.

Here is how the actions work with this implementation:
1. The agent starts the apps (osquerybeat in particular)
2. The osquerybeat receives the configuration from the agent
3. The osquerybeat parses configuration and sends the supported input types and osqueryd version back to the agent as a payload with the status update.  Along with the payload it also sends the version of osqueryd binary that is shipped with the agent. The requirement for the product is to be able to surface this information to the end-users. 
4. The osquerybeat also registers the action handlers with the rpc client for it's input types (in this case the input type is "osquery")
5. The agent received the input types in the status update payload and updates inputTypes set of the corresponding ApplicationState
6. When the new action is received from the Fleet Server, the Agent finds the matching ApplicationState for the input type and dispatches the action. 
7. The result of the action is forwarded back to Fleet Server. The query result data is send directly to elasticsearch by osquerybeat.


#### Packaging:
The osquerybeat build downloads the official distro from https://osquery.io/downloads/official and caches it locally for incremental builds.
The official distro binaries are provided as:
* .tar.gz for linux
* .pkg for MacOS
* .msi for Winderz

For osquerybeat package the osqueryd binary is extracted from the official distro(the only thing that we need) for the linux package. For windows and Mac the .pkg or .msi file is included into the package.
When osquerybeat if ran for the first time, depending on the platform it either unpacks the the osqueryd from the distro on that platform first or starts osqueryd as a child process right away. 

#### Disclaimer:
This is my first attempt on digging through the agent code, so there is a possibility that the lines of responsibilities for different pieces of the code are not exactly where they intended to be as designed by the original contributors. Would need some feedback.
Also not sure I covered all the bases with CI build for osquerybeat to get it correctly published to artifactory etc.
Might need some advice this.
Fully expect to iterate on osquerybeat customization further after this.